refactor: combine rangepicker input with datepicker input
