+++ src/components/DateInput/DateInput.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { ChangeEvent } from 'react';
 
 import {
   CalendarIcon,
@@ -6,43 +6,37 @@ import {
   DateInputWrapper,
   Input,
 } from '@/components/DateInput/DateInput.styled';
-import { useDateInput } from '@/hooks/useDateInput';
 
 export interface DateInputProps {
-  currentDate: Date | null;
+  value: string;
+  onClearInput: () => void;
+  onChange: (event: ChangeEvent<HTMLInputElement>) => void;
   toggleCalendar: () => void;
-  onInputValue: (dateString: string) => void;
   isError: boolean;
 }
 
 export const DateInput = ({
-  currentDate,
+  value,
+  onClearInput,
+  onChange,
   toggleCalendar,
-  onInputValue,
   isError,
-}: DateInputProps) => {
-  const { value, onClearInput, onChange } = useDateInput(
-    onInputValue,
-    currentDate,
-  );
-
-  return (
-    <DateInputWrapper>
-      <CalendarIcon onClick={toggleCalendar} />
-      <Input
-        value={value}
-        onChange={onChange}
-        placeholder='Choose Date'
-        type='text'
-        style={{
-          borderColor: isError ? 'red' : '#ddd',
-        }}
-      />
-      {value.length > 0 && (
-        <CrossButton onClick={onClearInput} type='button'>
-          ✖
-        </CrossButton>
-      )}
-    </DateInputWrapper>
-  );
-};
+}: DateInputProps) => (
+  <DateInputWrapper>
+    <CalendarIcon onClick={toggleCalendar} />
+    <Input
+      value={value}
+      onChange={onChange}
+      placeholder='Choose Date'
+      type='text'
+      style={{
+        borderColor: isError ? 'red' : '#ddd',
+      }}
+    />
+    {value.length > 0 && (
+      <CrossButton onClick={onClearInput} type='button'>
+        ✖
+      </CrossButton>
+    )}
+  </DateInputWrapper>
+);
